Dump Jan 13

+++ gslam/primitives.py
@@ -66,7 +66,9 @@ class Pose(torch.nn.Module):
     def forward(self) -> torch.Tensor:
         if not self.is_learnable:
             return self.Rt
-        id = torch.tensor([1, 0, 0, 0, 1, 0], device=self.Rt.device)
+        id = torch.tensor(
+            [1, 0, 0, 0, 1, 0], device=self.Rt.device, dtype=torch.float32
+        )
         rot = unvmap(rotation_6d_to_matrix)(self.dR + id)
         transform = torch.eye(4, device=self.Rt.device)
         transform[..., :3, :3] = rot
